Adding Westend genesis.
We are working on projects that utilize this package, In our projects we need to have Westend testnet, to let the community to try and learn Polkadot ecosystem.

Sometimes Developers have issues adding this manually as asked here: w3f/Grant-Milestone-Delivery#387 (comment),
hence merging this pull request will help a lot.

Thank you.